添加全局缓存，完成RentRepository

+++ BookRent/Data/BookRent.sql
@@ -16,8 +16,8 @@ create table Persons (
 );
 
 create table Rents (
-	PersonId int,
-	BookId int,
+	PersonId long,
+	BookId long,
 	StartDate date,
 	EndDate date
 );
